Move mapping utils to seperate directory
